refactor to exceptions and subcommands
