--- docs/Common.md
@@ -1,2 +1,0 @@
-# Common
-::: mininterface.common
