+++ docs/Subcommands.md
@@ -1,0 +1,2 @@
+# Subcommands
+::: mininterface.subcommands
