+++ mininterface/ValidationFail.py
@@ -1,0 +1,5 @@
+class ValidationFail(ValueError):
+    """ Signal to the form that submit failed and we want to restore it.
+    """
+    # NOTE example
+    pass
