+++ mininterface/subcommands.py
@@ -1,4 +1,4 @@
-""" Common helper objects that might make sense to be used outside the library. """
+""" Dealing with CLI subcommands, `from mininterface.subcommands import *`  """
 from abc import ABC, abstractmethod
 from dataclasses import dataclass
 from typing import TYPE_CHECKING
@@ -7,8 +7,6 @@ if TYPE_CHECKING:
     from .facet import Facet
 else:
     Facet = None
-
-# TODO create an exceptions.py
 
 
 @dataclass
@@ -20,7 +18,7 @@ class Command(ABC):
 
 
     What if I TODO
-    Special placeholder class Subcommand.
+    Special placeholder class SubcommandPlaceholder.
     This special class let the user to choose the subcommands via UI,
     while still benefiniting from default CLI arguments.
 
@@ -39,6 +37,8 @@ class Command(ABC):
     An example of Command usage:
 
     ```python
+    from mininterface.subcommands import Command
+
     @dataclass
     class SharedArgs:
         common: int
@@ -75,8 +75,8 @@ class Command(ABC):
 
     def init(self):
         """ Just before the form appears.
-        As the __post_init__ method is guaranteed to run just once (internal CLI behaviour),
-        you are welcome to override this method instead. You can use [self._facet][mininterface.Facet] from within.
+        As the `__post_init__` method is not guaranteed to run just once (internal CLI behaviour),
+        you are welcome to override this method instead. You can use [self._facet][mininterface.facet.Facet] from within.
         """
         ...
 
@@ -95,32 +95,3 @@ class SubcommandPlaceholder(Command):
 
 
 SubcommandPlaceholder.__name__ = "subcommand"  # show just the shortcut in the CLI
-
-
-class Cancelled(SystemExit):
-    """ User has cancelled.
-    A SystemExit based exception noting that the program exits without a traceback,
-    ex. if user hits the escape or closes the window. """
-    # We inherit from SystemExit so that the program exits without a traceback on ex. GUI escape.
-    pass
-
-
-class ValidationFail(ValueError):
-    """ Signal to the form that submit failed and we want to restore it.
-    """
-    # NOTE example
-    pass
-
-
-class DependencyRequired(ImportError):
-    def __init__(self, extras_name):
-        super().__init__(extras_name)
-        self.message = extras_name
-
-    def __str__(self):
-        return f"Required dependency. Run: pip install mininterface[{self.message}]"
-
-
-class InterfaceNotAvailable(ImportError):
-    """ Interface failed to init, ex. display not available in GUI. Or the underlying dependency was uninstalled. """
-    pass
