+++ mininterface/tk_interface/tk_window.py
@@ -6,9 +6,11 @@ from tktooltip import ToolTip
 
 from tkinter_form import Form, Value
 
+from ..exceptions import Cancelled
+
 from ..facet import BackendAdaptor
 from ..form_dict import TagDict, formdict_to_widgetdict
-from ..common import Cancelled, ValidationFail
+from ..ValidationFail import ValidationFail
 from ..tag import Tag
 from .tk_facet import TkFacet
 from .utils import recursive_set_focus, replace_widgets
